Restructured the client folders, modified prison with start script , and web pack for css-loaders
