+++ package.json
@@ -6,13 +6,14 @@
   "author": "Austin <elektricwebdesign@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "dev": "NODE_ENV=development yarn start:server | yarn start:react",
+    "start": "pm2 start ecosystem.config.js",
     "start:react": "NODE_ENV=development webpack-dev-server --config webpack.dev.config --open ",
     "start:server": "pm2 start ecosystem.config.js",
+    "dev": "NODE_ENV=development yarn start:server | yarn start:react",
     "webpack": "webpack --config webpack.dev.config.js ",
     "test": "jest",
     "test:watchAll": "jest --watch",
-    "build:prod": "NODE_ENV=production webpack --config webpack.prod.config.js"
+    "build:prod": "NODE_ENV=production webpack --config webpack.prod.config.js -p"
   },
   "babel": {
     "presets": [
@@ -77,6 +78,7 @@
     "eslint-plugin-react": "^7.3.0",
     "jest": "^21.1.0",
     "raw-loader": "^0.5.1",
+    "style-loader": "^0.18.2",
     "uglifyjs-webpack-plugin": "^0.4.6",
     "webpack-dev-middleware": "^1.12.0",
     "webpack-hot-middleware": "^2.19.1",
