+++ server/server.js
@@ -10,6 +10,13 @@ const PROD = process.env.NODE_ENV === 'production';
 app.set('view engine', 'ejs');
 app.set('views', path.join(__dirname, 'src/views'));
 app.use(express.static(path.join(__dirname, '../build/dist')));
+app.use('/test', (req, res) => {
+  res.send({ Hello: 'uTile is Served' });
+});
+/**
+NOTE: We have to place all middleware that will handle all api requests above our get('*')
+There is probably a better way to do this but for now this will do
+*/
 if (PROD) {
   app.get('/', async (req, res) => {
 
@@ -18,14 +25,8 @@ if (PROD) {
     res.render('index', { initalContent });
   });
 }
-app.get('/', (req, res) => {
+app.get('*', (req, res) => {
   res.render('index', { initalContent: 'Loading' });
-})
-app.get('/test', (req, res) => {
-  res.json({ Hello: 'uTile FUCKING ROCKS' });
-});
-app.get('/fuckAll', (req, res) => {
-  res.json({ Fuck: 'Hellyea Bitchessss' });
 });
 // Serve the files on port 3000.
 app.listen(config.port, () => {
