+++ webpack.dev.config.js
@@ -54,6 +54,22 @@ module.exports = {
         },
       },
     },
+    {
+      test: /\.css$/,
+      use: [
+        'style-loader',
+        {
+          loader: 'css-loader',
+          options: {
+            modules: true, // default is false
+            sourceMap: true,
+            importLoaders: 1,
+            localIdentName: '[name]--[local]--[hash:base64:8]',
+          },
+        },
+        'postcss-loader',
+      ],
+    },
     ],
   },
   plugins: [
